+++ tomcat/src/main/java/org/apache/coyote/domain/request/RequestBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.domain;
+package org.apache.coyote.domain.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
