+++ tomcat/src/main/java/org/apache/coyote/domain/response/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.domain;
+package org.apache.coyote.domain.response;
 
 import java.util.Arrays;
