+++ tomcat/src/main/java/org/apache/coyote/handler/Handler.java
@@ -2,8 +2,8 @@ package org.apache.coyote.handler;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
-import org.apache.coyote.domain.HttpRequest;
-import org.apache.coyote.domain.MyHttpResponse;
+import org.apache.coyote.domain.request.HttpRequest;
+import org.apache.coyote.domain.response.MyHttpResponse;
 
 public interface Handler {
