+++ tomcat/src/main/java/org/apache/coyote/handler/LoginHandler.java
@@ -6,11 +6,11 @@ import java.util.Optional;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.domain.FilePath;
-import org.apache.coyote.domain.HttpMethod;
-import org.apache.coyote.domain.HttpRequest;
-import org.apache.coyote.domain.HttpStatusCode;
-import org.apache.coyote.domain.MyHttpResponse;
-import org.apache.coyote.domain.RedirectUrl;
+import org.apache.coyote.domain.request.HttpMethod;
+import org.apache.coyote.domain.request.HttpRequest;
+import org.apache.coyote.domain.response.HttpStatusCode;
+import org.apache.coyote.domain.response.MyHttpResponse;
+import org.apache.coyote.domain.response.RedirectUrl;
 import org.apache.coyote.http11.Http11Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
