+++ tomcat/src/main/java/org/apache/coyote/handler/StaticFileHandler.java
@@ -3,9 +3,9 @@ package org.apache.coyote.handler;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import org.apache.coyote.domain.FilePath;
-import org.apache.coyote.domain.HttpRequest;
-import org.apache.coyote.domain.HttpStatusCode;
-import org.apache.coyote.domain.MyHttpResponse;
+import org.apache.coyote.domain.request.HttpRequest;
+import org.apache.coyote.domain.response.HttpStatusCode;
+import org.apache.coyote.domain.response.MyHttpResponse;
 
 public class StaticFileHandler implements Handler {
     @Override
