+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,18 +7,10 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.Socket;
 import java.net.URISyntaxException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
-import nextstep.jwp.model.User;
 import org.apache.coyote.Processor;
-import org.apache.coyote.domain.FilePath;
-import org.apache.coyote.domain.HttpRequest;
-import org.apache.coyote.domain.HttpStatusCode;
-import org.apache.coyote.domain.MyHttpResponse;
-import org.apache.coyote.domain.RedirectUrl;
+import org.apache.coyote.domain.request.HttpRequest;
+import org.apache.coyote.domain.response.MyHttpResponse;
 import org.apache.coyote.handler.HandlerMapping;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
